requirements.txt
streamlit==1.48.1
langchain==0.3.27
langchain-requests==0.1.4
python-dotenv==1.1.1
tabulate==0.9.0
openai

